Move Config to Theme namespace and NoTimber to Utils
